MR comments resolved.
