+++ RocketApp/Domain/Package.swift
@@ -15,6 +15,7 @@ let package = Package(
   ],
 
   dependencies: [
+    .package(path: "../Infrastructure"),
     .package(url: "https://github.com/Qase/swift-core", branch: "develop"),
     .package(url: "https://github.com/pointfreeco/swift-dependencies", from: "0.1.4"),
     .package(url: "https://github.com/pointfreeco/xctest-dynamic-overlay", from: "0.3.2")
@@ -24,6 +25,7 @@ let package = Package(
     .target(
       name: "RocketsClient",
       dependencies: [
+        .product(name: "NetworkClientExtensions", package: "Infrastructure"),
         .product(
           name: "Networking",
           package: "swift-core",
