+++ RocketApp/Domain/README.md
@@ -1,3 +1,7 @@
 # Domain
 
-A description of this package.
+Intermediate layer in dependency tree. Package contains domain tools, which are used for features but are depended 
+on low level tools in Infrastructure or external packages.
+
+Content:
+- Rocket client: tool for getting the data from API SpaceX, that are shown in Rocket detail and Rocket list
