+++ RocketApp/Domain/Sources/RocketsClient/RocketClient+liveValue.swift
@@ -1,5 +1,6 @@
 import Dependencies
 import Foundation
+import NetworkClientExtensions
 import Networking
 import RequestBuilder
