+++ RocketApp/Domain/Sources/RocketsClient/RocketsClient.swift
@@ -1,4 +1,5 @@
 import Combine
+import Dependencies
 import Foundation
 import Networking
 
@@ -8,4 +9,18 @@ public struct RocketsClient {
 
   public var getRocket: GetRocketFunction
   public var getAllRockets: GetAllRocketsFunction
+}
+
+enum RocketsClientKey: DependencyKey {
+  public static var liveValue: RocketsClient = .liveValue
+  #if DEBUG
+  public static var testValue: RocketsClient = .testValue
+  #endif
+}
+
+extension DependencyValues {
+  var rocketsClient: RocketsClient {
+    get { self[RocketsClientKey.self] }
+    set { self[RocketsClientKey.self] = newValue }
+  }
 }
