+++ RocketApp/Infrastructure/Package.swift
@@ -14,6 +14,10 @@ let package = Package(
       targets: ["CoreToolkit"]
     ),
     .library(
+      name: "NetworkClientExtensions",
+      targets: ["NetworkClientExtensions"]
+    ),
+    .library(
       name: "TCAExtensions",
       targets: ["TCAExtensions"]
     ),
@@ -24,7 +28,10 @@ let package = Package(
   ],
 
   dependencies: [
-    .package(url: "https://github.com/pointfreeco/swift-composable-architecture", .upToNextMajor(from: "0.0.0"))
+    .package(url: "https://github.com/Qase/swift-core", branch: "develop"),
+    .package(url: "https://github.com/pointfreeco/swift-dependencies", from: "0.1.4"),
+    .package(url: "https://github.com/pointfreeco/swift-composable-architecture", .upToNextMajor(from: "0.0.0")),
+    .package(url: "https://github.com/pointfreeco/xctest-dynamic-overlay", from: "0.3.2")
   ],
 
   targets: [
@@ -36,6 +43,14 @@ let package = Package(
     .testTarget(
       name: "CoreToolkitTests",
       dependencies: ["CoreToolkit"]
+    ),
+    .target(
+      name: "NetworkClientExtensions",
+      dependencies: [
+        .product(name: "Networking", package: "swift-core"),
+        .product(name: "Dependencies", package: "swift-dependencies"),
+        .product(name: "XCTestDynamicOverlay", package: "xctest-dynamic-overlay")
+      ]
     ),
     .target(
       name: "TCAExtensions",
