+++ RocketApp/Infrastructure/Sources/NetworkClientExtensions/NetworkClient+.swift
